+++ kirksdk/account_client.go
@@ -16,6 +16,7 @@ const appVersionPrefix = "/v3"
 var ErrInvalidAppURI = errors.New("app uri is invalid")
 
 type accountClientImp struct {
+	config    AccountConfig
 	accessKey string
 	secretKey string
 	host      string
@@ -27,6 +28,7 @@ type accountClientImp struct {
 func NewAccountClient(cfg AccountConfig) AccountClient {
 
 	p := new(accountClientImp)
+	p.config = cfg
 	p.host = cleanHost(cfg.Host)
 	p.transport = cfg.Transport
 	p.userAgent = cfg.UserAgent
@@ -42,6 +44,10 @@ func NewAccountClient(cfg AccountConfig) AccountClient {
 	}
 
 	return p
+}
+
+func (p *accountClientImp) GetConfig() (ret AccountConfig) {
+	return p.config
 }
 
 func (p *accountClientImp) GetAccountInfo(ctx context.Context) (ret AccountInfo, err error) {
@@ -173,6 +179,42 @@ func (p *accountClientImp) GetGrantedAppKey(ctx context.Context, appURI string) 
 	return
 }
 
+func (p *accountClientImp) GetAppspecs(ctx context.Context, specURI string) (ret SpecInfo, err error) {
+	url := fmt.Sprintf("%s%s/appspecs/%s", p.host, appVersionPrefix, specURI)
+	err = p.client.Call(ctx, &ret, "GET", url)
+	return
+}
+
+func (p *accountClientImp) ListPublicspecs(ctx context.Context) (ret []SpecInfo, err error) {
+	url := fmt.Sprintf("%s%s/publicspecs", p.host, appVersionPrefix)
+	err = p.client.Call(ctx, &ret, "GET", url)
+	return
+}
+
+func (p *accountClientImp) ListGrantedspecs(ctx context.Context) (ret []SpecInfo, err error) {
+	url := fmt.Sprintf("%s%s/grantedspecs", p.host, appVersionPrefix)
+	err = p.client.Call(ctx, &ret, "GET", url)
+	return
+}
+
+func (p *accountClientImp) GetVendorManagedAppStatus(ctx context.Context, appURI string) (ret VendorManagedAppStatus, err error) {
+	url := fmt.Sprintf("%s%s/apps/%s/status", p.host, appVersionPrefix, appURI)
+	err = p.client.Call(ctx, &ret, "PUT", url)
+	return
+}
+
+func (p *accountClientImp) GetVendorManagedAppEntry(ctx context.Context, appURI string) (ret VendorManagedAppEntry, err error) {
+	url := fmt.Sprintf("%s%s/apps/%s/entry", p.host, appVersionPrefix, appURI)
+	err = p.client.Call(ctx, &ret, "PUT", url)
+	return
+}
+
+func (p *accountClientImp) VendorManagedAppRepair(ctx context.Context, appURI string) (err error) {
+	url := fmt.Sprintf("%s%s/apps/%s/repair", p.host, appVersionPrefix, appURI)
+	err = p.client.Call(ctx, nil, "PUT", url)
+	return
+}
+
 func (p *accountClientImp) GetIndexClient(ctx context.Context) (client IndexClient, err error) {
 	accountInfo, err := p.GetAccountInfo(ctx)
 	if err != nil {
@@ -259,7 +301,7 @@ func (p *accountClientImp) GetQcosClient(ctx context.Context, appURI string) (cl
 	go func() {
 		result := getKeyFunc()
 
-		if result.ak == "" {
+		if result.ak == "" && result.err == nil {
 			result.err = fmt.Errorf("Fail to find keys for app \"%s\"", appURI)
 		}