+++ kirksdk/qcos_api.go
@@ -9,6 +9,8 @@ import (
 )
 
 type QcosClient interface {
+	GetConfig() (ret QcosConfig)
+
 	// GET /v3/stacks
 	ListStacks(ctx context.Context) (ret []StackInfo, err error)
 
@@ -333,6 +335,9 @@ type QcosClient interface {
 
 	// DELETE /v3/configservices/<namespace>
 	DeleteConfigServiceSpec(ctx context.Context, namespace string) (err error)
+
+	// POST /v3/webproxy
+	GetWebProxy(ctx context.Context, args GetWebProxyArgs) (ret WebProxyInfo, err error)
 }
 
 const (
@@ -749,12 +754,38 @@ type LogsSearchResult struct {
 }
 
 type Hit struct {
-	Log         string    `json:"log"`
-	CollectedAt time.Time `json:"collectedAt"`
-	PodIP       string    `json:"podIp"`
-	ProcessName string    `json:"processName"`
-	GateID      string    `json:"gateId"`
-	Domain      string    `json:"domain"`
+	CollectedAt     time.Time `json:"collectedAt" repo:"pod,access"`
+	CollectedAtNano int64     `json:"collectedAtNano"`
+	Host            string    `json:"host"`
+
+	Log           string `json:"log" repo:"pod"`
+	Path          string `json:"path" repo:"pod"`
+	Pattern       string `json:"pattern" repo:"pod"`
+	ContainerId   string `json:"containerId"`
+	ContainerName string `json:"containerName"`
+	JobInstance   string `json:"jobInstance" repo:"pod"`
+	JobTask       string `json:"jobTask" repo:"pod"`
+	PodIp         string `json:"podIp" repo:"pod"`
+	PodName       string `json:"podName" repo:"pod"`
+	PodVer        string `json:"podVer"`
+	ProcessName   string `json:"processName"`
+	Sip           string `json:"sip"`
+	Source        string `json:"source" repo:"pod"`
+
+	Type           string    `json:"type" repo:"access"`
+	RequestApp     string    `json:"requestApp" repo:"access"`
+	GateId         string    `json:"gateId" repo:"access"`
+	StartAt        time.Time `json:"startAt" repo:"access"`
+	Method         string    `json:"method" repo:"access"`
+	Url            string    `json:"url" repo:"access"`
+	ReqId          string    `json:"reqId" repo:"access"`
+	StatusCode     int64     `json:"statusCode" repo:"access"`
+	ElapsedNano    int64     `json:"elapsedNano" repo:"access"`
+	RequestHeader  string    `json:"requestHeader" repo:"access"`
+	RequestParams  string    `json:"requestParams" repo:"access"`
+	RequestBody    string    `json:"requestBody" repo:"access"`
+	ResponseHeader string    `json:"responseHeader" repo:"access"`
+	ResponseBody   string    `json:"responseBody" repo:"access"`
 }
 
 var (
@@ -929,4 +960,13 @@ type ConfigServiceSpecInfo CreateConfigServiceSpecArgs
 type UpdateConfigServiceSpecArgs struct {
 	Vars     map[string]interface{}   `json:"vars"`
 	Listvars []map[string]interface{} `json:"listvars"`
+}
+
+type GetWebProxyArgs struct {
+	Backend string `json:"backend"`
+}
+
+type WebProxyInfo struct {
+	Backend    string `json:"backend"`
+	OneTimeURL string `json:"oneTimeUrl"`
 }